requirements.txt
Flask
werkzeug
selenium
python-dotenv
google-generativeai
